refactor: domain, service, view 패키지 추가 및 구조 변경
