+++ src/main/java/racingcar/RacingGameApp.java
@@ -1,5 +1,11 @@
 package racingcar;
 
+import racingcar.domain.RacingCar;
+import racingcar.service.RacingGame;
+import racingcar.service.RacingGameReferee;
+import racingcar.view.RacingGameInputView;
+import racingcar.view.RacingGameScoreView;
+
 import java.util.ArrayList;
 import java.util.List;
