+++ src/main/java/racingcar/service/RandomMovementManager.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.service;
 
 public class RandomMovementManager {
