+++ src/main/java/racingcar/service/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.service;
 
 @FunctionalInterface
 public interface RandomNumberGenerator {
