+++ src/main/java/racingcar/view/RacingGameInputView.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.view;
+
+import racingcar.domain.RacingCar;
 
 import java.util.Arrays;
 import java.util.List;
