+++ src/test/java/racingcar/service/RacingGameTest.java
@@ -1,8 +1,9 @@
-package racingcar;
+package racingcar.service;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.RacingCar;
 
 import java.util.ArrayList;
 import java.util.Arrays;
