+++ src/test/java/racingcar/service/RandomMovementManagerTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.service;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
