+++ src/test/java/racingcar/view/RacingGameInputViewTest.java
@@ -1,10 +1,11 @@
-package racingcar;
+package racingcar.view;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.domain.RacingCar;
 
 import java.io.ByteArrayInputStream;
 import java.util.List;
