+++ src/test/java/racingcar/view/RacingGameScoreViewTest.java
@@ -1,8 +1,9 @@
-package racingcar;
+package racingcar.view;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.RacingCar;
 
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
